// services/server/src/modules/tech/spider/tasks/interfaces/mayigeek.interface.ts
export interface MayigeekTask {
  id: string //"1258"
  name: string //"企业数字化服务平台项目"

  desc: string //"工业行业数字化互联平台开发。实现工业行业供应商、主管部门，采购企业间的互联互通。包括问卷、评估诊断、跨平台数据采集互通、报表等功能。\r\n"
  skill_desc: string //"技术能力合格，诚信有耐心！"

  min_price: number //50000

  day: string //"60"
  rest_day: number //0，剩余天数

  apply_times: string // 申请此处
  browse_times: string //"378"，查看次数
  start: string //"0000-00-00"
  end: string //"0000-00-00"

  bidStatus: string //"选标中"，当前状态
  bid_status: string //"1"
  bid_end: string //"2021-11-01"，结束时间

  deadline: string //"2021-12-01"

  create_date: string //"2021-10-13 08:14:58"，创建时间

  status: string //"6"
  type: string //""
  typeStr: string //""
}